feat: Update to Vue 3.0 and Vue Test Utils 2.0 (#137) (#163)
BREAKING CHANGE:
- Snapshots are potentially different due to whitespaces.
- Some options such as stubs are now moved to `global.stubs`.
- Previous usage of vue instance as a parameter for the callback function is not replaced with global `config` options (for example, `global.directives`).
- `isUnmounted` is gone.
- `updateProps` is now called setProps to align with VTU.

Some of the missing steps is to provide support to Vue Router, and also to update libraries of the ecosystem to their Vue 3 version.

This is an exciting release! 🎉  It marks the first release aiming to support Vue 3 and Vue Test Utils 2. Please [head to the PR](https://github.com/testing-library/vue-testing-library/pull/137) to get more information, and feel free to open up issues and PRs to fix missing features / ports 😄